Fix processing flow bug
The use of l.IsUpdated() to determine if processing should continue
turns out to be flawed because the updated flag is only set if the
status is changed and if the status is already what it should be
it is not updated. The updated flag is meant to be used to determine
if the update function should be called at the end of the reconcilation
so using it for this purpose is effectively a misuse.

As it turns out the delayed flag is set if an action is taken that
means the processing should stop and the layer should be requeue after
a delay so we can safely use this flag to suppress further processing.

However this too is a misuse. We should probably add a done flag that
is set when we want the reconcilation processing to cease?

For now I've made this simple change to fix the bug in the short term.