Fix Setup Script for corporate network use
Changes to setup script to facilitate the deployment of Kraan and
GitOps toolkit components in a corprate network. Includes a fix
to the code that added proxy related environment variables to the
GitOps toolkit source controller so it can access external git
repositories.

Also added a --no-git-auth option to explictly remove the ref-secret
from the GitRepostitory custom resource included in the testdata.