Update processing of gitrepository syncronization
This change improves the processing of gitrepository objects which the
controller watches. Now it will only retrieve data from the source
controller for a gitrepository that an addonslayer references.

Also improved mutex handling so we don't block addonslayer processing
while the data is being fetched from the source controller and we
fetch to a separate directory then do an remove/rename to move the
new data into place as quickly as possible.

Finally the reconcile processing now checks if the repository object
is present and waits for up to 15 seconds for it to be created before
updating the symbolic link to the data.