Improve GitRepository Sync logic
Currently the kraan controller syncs the GitRepositories used by
addonslayers every time the GitReposity custom resource is updated.

This change as a check to see if the revision field in the status
of the GitRepository has changed and only proceeds with if a sync
and requeue of the addonslayers using it if revision has changed.