Add check for GitRepository ready
Added a check to the dependencies check and reconcile processing to
verify that the source GitRepository is ready. The Source Controller
does not reset the last good artifact details in the status when the
spec is changed so it is necessary to verify that the git repository
has been synced correctly.

When the Kraan-Controller processes GitRepository custom resources a log
message causes a panic if the GitRepository has no Status.Artifact
element. Created a helper function to cater for this.

Also move functions used by multiple packages to a utils package.