Expose kubectl error in layer status
It has been noted that when the kubectl apply fails no information is
available to the operator relating to the reason for the failure.

This patch addresses that by exposing the kubectl stderr in the status
condition message.

Also added function and source/line to the error message which may be
useful for issue reporting in the case where the cause is a kraan bug
rather than user manifests or cluster state.