Support updateVersion annotation (#150)
See https://github.com/fidelity/kraan/issues/148, this patch
implements a check of HelmRelease source definitions for an
annotation of kraan.updateVersion: true. If this is present
a value with key kraanVersion will be added to the chart
containing the current addonsLayer version. If this is
different from the verision set when the HelmRelease was
previously deployed it will cause the HelmRelease to be redeployed.